Develop initial implementation of bot
Built the initial implementation of this bot.

The bot is configured with config.yml. In that config file, the name of
the bot is specified, as well as command definitions.

There are currently only two types of commands supported: value and
reddit.

 - A 'value' command will just return params.value.
 - A 'reddit' command will get top 50 hot posts from params.sub and
   return a random one.

For authenticating with Discord and Reddit, a .env file must be placed
in the working directory with the following variables defined:

 - DISCORD_TOKEN
 - REDDIT_USER_AGENT
 - REDDIT_CLIENT_ID
 - REDDIT_CLIENT_SECRET
 - REDDIT_USERNAME
 - REDDIT_PASSWORD